pyproject.toml
[tool.poetry]
name = "cryptoguardian"
version = "0.1.0"
description = ""
authors = ["Sentiago <arseny317@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
aiosqlite = "0.20.0"
annotated-types = "0.6.0"
anyio = "4.3.0"
astroid = "3.1.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
bcrypt = "4.1.3"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "42.0.7"
dill = "0.3.8"
dnspython = "2.6.1"
email-validator = "2.1.1"
environs = "11.0.0"
exceptiongroup = "1.2.1"
fastapi = "0.111.0"
fastapi-cli = "0.0.3"
frozenlist = "1.4.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
isort = "5.13.2"
Jinja2 = "3.1.4"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
marshmallow = "3.21.2"
mccabe = "0.7.0"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy = "1.10.0"
mypy-extensions = "1.0.0"
orjson = "3.10.3"
packaging = "24.0"
platformdirs = "4.2.1"
pycparser = "2.22"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
Pygments = "2.18.0"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
pylint = "3.1.1"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
PyYAML = "6.0.1"
requests = "2.31.0"
rich = "13.7.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.30"}
starlette = "0.37.2"
tomli = "2.0.1"
tomlkit = "0.12.5"
typer = "0.12.3"
typing-extensions = "4.11.0"
ujson = "5.9.0"
urllib3 = "2.2.1"
uvicorn = "0.29.0"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"
yarl = "1.9.4"
zxcvbn = "4.4.28"
sqlmodel = "^0.0.18"
black = "^24.4.2"
pydantic-settings = "^2.2.1"
sentry-sdk = "^2.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
